fix: 修复所有隐式类型截断的编译警告

+++ src/runtime/interp/basic_block_splitter.h
@@ -17,18 +17,18 @@ class BasicBlockSplitter
 
     RtResultVoid split();
 
-    const utils::HashSet<size_t>& get_split_offsets() const;
+    const utils::HashSet<uint32_t>& get_split_offsets() const;
 
   private:
-    void mark_valid_il_offset(size_t offset);
-    bool is_valid_il_offset(size_t offset) const;
+    void mark_valid_il_offset(uint32_t offset);
+    bool is_valid_il_offset(uint32_t offset) const;
 
     RtResultVoid split_codes();
     void split_exception_clauses();
     bool validate_offsets() const;
 
     const metadata::RtMethodBody* _method_body;
-    utils::HashSet<size_t> _split_offsets;
+    utils::HashSet<uint32_t> _split_offsets;
     uint32_t* _valid_il_offsets;
     size_t _valid_il_offsets_count;
 };
